Vittali said its 70% ready
