+++ src/main/java/org/aerogear/plugin/intellij/mobile/ui/actions/ClientCreatedCheckAction.java
@@ -1,68 +1,37 @@
 package org.aerogear.plugin.intellij.mobile.ui.actions;
 
-
 import com.intellij.json.JsonFileType;
-import com.intellij.notification.NotificationListener;
-
+import com.intellij.openapi.actionSystem.AnAction;
+import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.command.WriteCommandAction;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.startup.StartupActivity;
-import com.intellij.openapi.vfs.LocalFileSystem;
-import com.intellij.openapi.vfs.VirtualFile;
 import org.aerogear.plugin.intellij.mobile.api.CLIRunnerImpl;
 import org.aerogear.plugin.intellij.mobile.api.MobileAPI;
 import org.aerogear.plugin.intellij.mobile.models.MobileClient;
-import org.aerogear.plugin.intellij.mobile.services.MobileNotificationsService;
-import org.aerogear.plugin.intellij.mobile.ui.MobileIcons;
+import org.aerogear.plugin.intellij.mobile.ui.createclientpopup.Constants;
 import org.aerogear.plugin.intellij.mobile.ui.createclientpopup.CreateClientForm;
 import org.aerogear.plugin.intellij.mobile.utils.WriteToFileRunnable;
-import org.jetbrains.annotations.NotNull;
 
-
-public class ClientCreatedCheckAction implements StartupActivity {
-    private static final String dotFileName = ".aerogear.json";
-
+public class ClientCreatedCheckAction extends AnAction {
     @Override
-    public void runActivity(@NotNull Project project) {
-        String projectPath = project.getBasePath();
-        checkFile(projectPath + '/' + dotFileName, project);
+    public void actionPerformed(AnActionEvent e) {
+        MobileAPI mobileAPI = new MobileAPI(new CLIRunnerImpl());
+        String filePath = e.getProject().getBasePath() + "/" + Constants.DOT_FILENAME;
+        showCreateClientForm(e.getProject(), mobileAPI, filePath);
     }
 
+    public void showCreateClientForm(Project project, MobileAPI mobileAPI, String filePath) {
+        CreateClientForm clientForm = new CreateClientForm(project, mobileAPI);
+        clientForm.show();
 
-    public void checkFile(String filePath, @NotNull Project project) {
-        VirtualFile file = LocalFileSystem.getInstance().refreshAndFindFileByPath(filePath);
-
-        if (file == null) {
-            MobileAPI mobileAPI = new MobileAPI(new CLIRunnerImpl());
-
-            NotificationListener notificationListener = (notification, event) -> {
-                {
-                    CreateClientForm clientForm = new CreateClientForm(project, mobileAPI);
-                    clientForm.show();
-
-                    if (CreateClientForm.OK_EXIT_CODE == clientForm.getExitCode()) {
-                        MobileClient mobileClient = mobileAPI.createClient(clientForm.getName(), clientForm.getClientType(), clientForm.getAppId()
-                        );
-
-                        CharSequence cs = mobileClient.getSpec().toJsonPrettyPrint();
-
-                        WriteToFileRunnable writeToFile = new WriteToFileRunnable(project, filePath, cs, dotFileName, JsonFileType.INSTANCE);
-                        WriteCommandAction.runWriteCommandAction(project, writeToFile);
-                    }
-                }
-            };
-
-            new MobileNotificationsService().notifyWarning(
-                    MobileIcons.AEROGEAR,
-                    "<html>Client app definition not found",
-                    " <a href=\"" + "here" + "\" target=\"blank\">Create new client app definition</a> </html>",
-                    notificationListener
+        if (CreateClientForm.OK_EXIT_CODE == clientForm.getExitCode()) {
+            MobileClient mobileClient = mobileAPI.createClient(clientForm.getName(), clientForm.getClientType(), clientForm.getAppId()
             );
 
+            CharSequence cs = mobileClient.getSpec().toJsonPrettyPrint();
+
+            WriteToFileRunnable writeToFile = new WriteToFileRunnable(project, filePath, cs, Constants.DOT_FILENAME, JsonFileType.INSTANCE);
+            WriteCommandAction.runWriteCommandAction(project, writeToFile);
         }
-
-
     }
-
-
 }
