+++ src/main/java/org/aerogear/plugin/intellij/mobile/ui/createclientpopup/Constants.java
@@ -1,9 +1,11 @@
 package org.aerogear.plugin.intellij.mobile.ui.createclientpopup;
 
-class Constants {
+public class Constants {
     public static final String[] CLIENT_TYPES = new String[]{
             "android",
             "cordova",
             "iOS"
     };
+
+    public static String DOT_FILENAME = ".aerogear.json";
 }
